requirements.txt
ultralytics
PyYAML
opencv-python
numpy